DatabaseInit completed

+++ TransmiApp/TransmiApp/src/main/java/com/example/TransmiApp/repository/AssignmentRepository.java
@@ -1,10 +1,11 @@
 package com.example.TransmiApp.repository;
 
-import com.example.TransmiApp.model.Assignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.TransmiApp.model.Assignment;
     
 @Repository
-public interface AssignmentRepository extends JpaRepository<Assignment, Integer> {
+public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
     
 }
