+++ TransmiApp/TransmiApp/src/main/java/com/example/TransmiApp/repository/BusRepository.java
@@ -1,10 +1,11 @@
 package com.example.TransmiApp.repository;
 
-import com.example.TransmiApp.model.Bus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.TransmiApp.model.Bus;
     
 @Repository
-public interface BusRepository extends JpaRepository<Bus, Integer> {
+public interface BusRepository extends JpaRepository<Bus, Long> {
     
 }
