+++ TransmiApp/TransmiApp/src/main/java/com/example/TransmiApp/repository/DriverRepository.java
@@ -1,10 +1,11 @@
 package com.example.TransmiApp.repository;
 
-import com.example.TransmiApp.model.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.TransmiApp.model.Driver;
     
 @Repository
-public interface DriverRepository extends JpaRepository<Driver, Integer> {
+public interface DriverRepository extends JpaRepository<Driver, Long> {
     
 }
